feat: Enhance VoiceManager leaveVoice method to handle non-guild messages
feat: Add new GIF asset for GitHub cron job

chore: Update config with new owner and website links, change default prefix to 'C!', and update command descriptions

fix: Update ownerHelperFirewall to correctly respond with ephemeral messages

refactor: Improve command handling for setbalance, resetinv, and kick commands with better error messages

refactor: Simplify spamsendto and spamsay commands by delegating to discordFormat functions

feat: Implement nanamiHelpMenu and nanamiBotInfo functions for better command organization

fix: Update give and take money commands to handle user balances more effectively

chore: Rename backupfiles command to backup for clarity

+++ ClassFunction/VoiceManager.js
@@ -985,6 +985,7 @@ class VoiceManager {
     // Method untuk leave voice channel
     async leaveVoice(message) {
       const guildId = message.guild.id;
+      if(!guildId) return message.reply({content: `${discordEmotes.error} You are not in a server!`, ephemeral: true});
       const connection = useQueue(guildId);
       if (connection) {
         connection.delete();
